Work on the PaymentController

+++ src/main/java/com/davy/restapi/orderlines/dto/OrderLineDetail.java
@@ -4,6 +4,8 @@ import com.davy.restapi.product.dto.ProductDTO;
 import com.fasterxml.jackson.annotation.JsonProperty;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OrderLineDetail {
 
@@ -17,9 +19,12 @@ public class OrderLineDetail {
     private short quantity;
 
     @JsonProperty("totalPrice")
-    private float totalPrice;
+    private BigDecimal totalPrice;
 
-    public OrderLineDetail(Long id, ProductDTO product, short quantity, float totalPrice) {
+    public OrderLineDetail(Long id,
+                           ProductDTO product,
+                           short quantity,
+                           BigDecimal totalPrice) {
         this.id = id;
         this.product = product;
         this.quantity = quantity;
