+++ src/main/java/com/davy/restapi/product/dto/ProductDTO.java
@@ -2,8 +2,10 @@ package com.davy.restapi.product.dto;
 
 import com.fasterxml.jackson.annotation.JsonProperty;
 
+import java.math.BigDecimal;
+
 public record ProductDTO(
         @JsonProperty("id")Long id,
         @JsonProperty("name")String name,
-        @JsonProperty("sellingPrice") float sellingPrice
+        @JsonProperty("sellingPrice") BigDecimal sellingPrice
 ){}
